requirements.txt
distributed~=2021.2.0
fastai~=2.2.5
fastcore~=1.3.19
pandas~=1.2.2
numpy~=1.19.2
pillow~=8.1.0
tqdm~=4.56.0